# artifacts/api-server/src/routes/admin.ts
import { Router, type IRouter } from "express";
import {
  CreateAdminLicenseKeyBody,
  CreateAdminLicenseKeyResponse,
  GetAdminPurchaseSettingsResponse,
  ListAdminLicenseKeysQueryParams,
  ListAdminLicenseKeysResponse,
  UpdateAdminPurchaseSettingsBody,
  UpdateAdminPurchaseSettingsResponse,
  RevokeAdminLicenseKeyParams,
  GetAdminOverviewResponse,
  ListAdminUsersQueryParams,
  ListAdminUsersResponse,
  GetAdminUserParams,
  GetAdminUserResponse,
  UpdateAdminUserSubscriptionParams,
  UpdateAdminUserSubscriptionBody,
  UpdateAdminUserSubscriptionResponse,
  GetAdminSystemSettingsResponse,
  UpdateAdminSystemSettingsBody,
  UpdateAdminSystemSettingsResponse,
  GetAdminOperationsResponse,
  UpdateAdminCampaignStatusParams,
  UpdateAdminCampaignStatusBody,
  UpdateAdminCampaignStatusResponse,
  RetryAdminCampaignTargetParams,
  RetryAdminCampaignTargetResponse,
} from "@workspace/api-zod";
import { and, desc, eq, inArray } from "drizzle-orm";
import {
  activityLogsTable,
  appUsersTable,
  campaignTargetsTable,
  campaignsTable,
  destinationsTable,
  proxiesTable,
  telegramAccountsTable,
  db,
} from "@workspace/db";
import {
  createAdminLicenseKeys,
  listAdminLicenseKeys,
  revokeAdminLicenseKey,
  getAdminOverview,
  listAdminUsers,
  getAdminUser,
  updateSubscriptionByAdmin,
} from "../lib/subscriptions";
import { requireAdmin } from "../middlewares/authMiddleware";
import { isTelegramPurchaseUrl, getPurchaseSettings, updatePurchaseSettings } from "../lib/purchase-settings";
import { recordActivity } from "../lib/activity";
import { getSystemSettings, updateSystemSettings } from "../lib/system-settings";

const router: IRouter = Router();

function sendError(res: any, status: number, error: string): void {
  res.status(status).json({ error });
}

router.use("/admin", requireAdmin);

router.get("/admin/overview", async (_req, res): Promise<void> => {
  res.json(GetAdminOverviewResponse.parse(await getAdminOverview()));
});

router.get("/admin/users", async (req, res): Promise<void> => {
  const parsed = ListAdminUsersQueryParams.safeParse(req.query);
  if (!parsed.success) {
    sendError(res, 400, "Bộ lọc người dùng không hợp lệ.");
    return;
  }
  res.json(ListAdminUsersResponse.parse(await listAdminUsers(parsed.data)));
});

router.get("/admin/users/:userId", async (req, res): Promise<void> => {
  const parsed = GetAdminUserParams.safeParse(req.params);
  if (!parsed.success) {
    sendError(res, 400, "Người dùng không hợp lệ.");
    return;
  }
  const user = await getAdminUser(parsed.data.userId);
  if (!user) {
    sendError(res, 404, "Không tìm thấy người dùng.");
    return;
  }
  res.json(GetAdminUserResponse.parse(user));
});

router.patch("/admin/users/:userId/subscription", async (req, res): Promise<void> => {
  const params = UpdateAdminUserSubscriptionParams.safeParse(req.params);
  const body = UpdateAdminUserSubscriptionBody.safeParse(req.body);
  if (!params.success || !body.success || !Number.isInteger(body.data.durationDays)) {
    sendError(res, 400, "Thông tin gói đăng ký không hợp lệ.");
    return;
  }
  const outcome = await updateSubscriptionByAdmin({
    userId: params.data.userId,
    adminUserId: req.userId!,
    plan: body.data.plan,
    durationDays: body.data.durationDays,
  });
  if (!outcome.ok && outcome.reason === "not_found") {
    sendError(res, 404, "Không tìm thấy người dùng.");
    return;
  }
  if (!outcome.ok && outcome.reason === "downgrade") {
    sendError(res, 409, "Không thể hạ cấp gói đăng ký.");
    return;
  }
  if (!outcome.ok) {
    sendError(res, 400, "Chỉ có thể nâng cấp hoặc gia hạn các gói trả phí.");
    return;
  }
  res.json(UpdateAdminUserSubscriptionResponse.parse(outcome.subscription));
});

router.get("/admin/purchase-settings", async (_req, res): Promise<void> => {
  res.json(GetAdminPurchaseSettingsResponse.parse(await getPurchaseSettings()));
});

router.patch("/admin/purchase-settings", async (req, res): Promise<void> => {
  const parsed = UpdateAdminPurchaseSettingsBody.safeParse(req.body);
  if (!parsed.success) {
    sendError(res, 400, "Link Telegram Bot không hợp lệ.");
    return;
  }

  const telegramPurchaseUrl = parsed.data.telegramPurchaseUrl.trim();
  if (!isTelegramPurchaseUrl(telegramPurchaseUrl)) {
    sendError(res, 400, "Link phải có dạng https://t.me/ten_bot hoặc https://telegram.me/ten_bot.");
    return;
  }

  const settings = await updatePurchaseSettings({
    telegramPurchaseUrl,
    updatedBy: req.userId!,
  });
  await recordActivity({
    ownerUserId: req.userId!,
    event: "purchase_settings.updated",
    message: "Updated Telegram purchase link",
    level: "success",
    metadata: { hostname: new URL(telegramPurchaseUrl).hostname },
  });
  res.json(UpdateAdminPurchaseSettingsResponse.parse(settings));
});

router.get("/admin/system-settings", async (_req, res): Promise<void> => {
  res.json(GetAdminSystemSettingsResponse.parse(await getSystemSettings()));
});

router.patch("/admin/system-settings", async (req, res): Promise<void> => {
  const parsed = UpdateAdminSystemSettingsBody.safeParse(req.body);
  if (!parsed.success) return void sendError(res, 400, "Cấu hình hệ thống không hợp lệ.");
  const settings = parsed.data;
  const allLimits = Object.values(settings.planLimits);
  const allIntegerLimits = allLimits.every((limit) => (
    [limit.accountLimit, limit.campaignLimit, limit.messageDailyLimit]
      .every((value) => value === null || Number.isInteger(value))
  ));
  const defaults = settings.campaignDefaults;
  const validDefaults = [
    defaults.maxRetries,
    defaults.delayMinSeconds,
    defaults.delayMaxSeconds,
    defaults.roundDelayMinSeconds,
    defaults.roundDelayMaxSeconds,
    settings.defaultAccountDailyLimit,
  ].every(Number.isInteger)
    && defaults.delayMinSeconds <= defaults.delayMaxSeconds
    && defaults.roundDelayMinSeconds <= defaults.roundDelayMaxSeconds;
  try {
    Intl.DateTimeFormat(undefined, { timeZone: settings.defaultTimezone });
  } catch {
    return void sendError(res, 400, "Múi giờ mặc định không hợp lệ.");
  }
  if (!allIntegerLimits || !validDefaults) return void sendError(res, 400, "Cấu hình giới hạn hoặc delay không hợp lệ.");
  const updated = await updateSystemSettings(settings, req.userId!);
  await recordActivity({
    ownerUserId: req.userId!,
    event: "system_settings.updated",
    level: "success",
    message: "Updated system settings",
    metadata: {
      registrationEnabled: updated.registrationEnabled,
      maintenanceMode: updated.maintenanceMode,
      defaultTimezone: updated.defaultTimezone,
    },
  });
  res.json(UpdateAdminSystemSettingsResponse.parse(updated));
});

router.get("/admin/operations", async (_req, res): Promise<void> => {
  const [users, accounts, proxies, campaigns, targets, logs] = await Promise.all([
    db.select({ id: appUsersTable.id, username: appUsersTable.username }).from(appUsersTable),
    db.select({
      id: telegramAccountsTable.id,
      ownerUserId: telegramAccountsTable.ownerUserId,
      name: telegramAccountsTable.name,
      status: telegramAccountsTable.status,
      proxyId: telegramAccountsTable.proxyId,
      lastSyncAt: telegramAccountsTable.lastSyncAt,
      cooldownUntil: telegramAccountsTable.cooldownUntil,
    }).from(telegramAccountsTable),
    db.select({ id: proxiesTable.id, name: proxiesTable.name, status: proxiesTable.status }).from(proxiesTable),
    db.select({
      id: campaignsTable.id,
      ownerUserId: campaignsTable.ownerUserId,
      name: campaignsTable.name,
      status: campaignsTable.status,
    }).from(campaignsTable).orderBy(desc(campaignsTable.createdAt)),
    db.select({
      id: campaignTargetsTable.id,
      campaignId: campaignTargetsTable.campaignId,
      destinationId: campaignTargetsTable.destinationId,
      status: campaignTargetsTable.status,
      attempts: campaignTargetsTable.attempts,
      nextAttemptAt: campaignTargetsTable.nextAttemptAt,
      lastError: campaignTargetsTable.lastError,
    }).from(campaignTargetsTable).orderBy(desc(campaignTargetsTable.updatedAt)),
    db.select({
      id: activityLogsTable.id,
      ownerUserId: activityLogsTable.ownerUserId,
      level: activityLogsTable.level,
      event: activityLogsTable.event,
      message: activityLogsTable.message,
      metadata: activityLogsTable.metadata,
      campaignId: activityLogsTable.campaignId,
      accountId: activityLogsTable.accountId,
      createdAt: activityLogsTable.createdAt,
    }).from(activityLogsTable).orderBy(desc(activityLogsTable.createdAt)).limit(200),
  ]);
  const destinations = await db.select({ id: destinationsTable.id, title: destinationsTable.title }).from(destinationsTable);
  const usernames = new Map(users.map((user) => [user.id, user.username]));
  const proxyById = new Map(proxies.map((proxy) => [proxy.id, proxy]));
  const destinationTitles = new Map(destinations.map((destination) => [destination.id, destination.title]));
  const targetsByCampaign = new Map<string, typeof targets>();
  for (const target of targets) targetsByCampaign.set(target.campaignId, [...(targetsByCampaign.get(target.campaignId) ?? []), target]);
  const campaignsById = new Map(campaigns.map((campaign) => [campaign.id, campaign]));

  res.json(GetAdminOperationsResponse.parse({
    accounts: accounts.map((account) => {
      const proxy = account.proxyId ? proxyById.get(account.proxyId) : undefined;
      const failedTargets = targets.filter((target) => target.status === "failed" && campaignsById.get(target.campaignId)?.ownerUserId === account.ownerUserId).length;
      return {
        id: account.id,
        ownerUsername: usernames.get(account.ownerUserId) ?? "Unknown",
        name: account.name,
        status: account.status,
        proxyName: proxy?.name ?? null,
        proxyStatus: proxy?.status ?? null,
        lastSyncAt: account.lastSyncAt,
        cooldownUntil: account.cooldownUntil,
        failedTargets,
      };
    }),
    campaigns: campaigns.slice(0, 100).map((campaign) => {
      const campaignTargets = targetsByCampaign.get(campaign.id) ?? [];
      return {
        id: campaign.id,
        ownerUsername: usernames.get(campaign.ownerUserId) ?? "Unknown",
        name: campaign.name,
        status: campaign.status,
        pendingTargets: campaignTargets.filter((target) => ["pending", "sending"].includes(target.status)).length,
        failedTargets: campaignTargets.filter((target) => target.status === "failed").length,
        sentTargets: campaignTargets.filter((target) => target.status === "sent").length,
      };
    }),
    targets: targets.filter((target) => ["pending", "sending", "failed", "requires_review"].includes(target.status)).slice(0, 200).map((target) => {
      const campaign = campaignsById.get(target.campaignId);
      return {
        id: target.id,
        campaignId: target.campaignId,
        campaignName: campaign?.name ?? "Unknown campaign",
        ownerUsername: campaign ? usernames.get(campaign.ownerUserId) ?? "Unknown" : "Unknown",
        destinationTitle: destinationTitles.get(target.destinationId) ?? null,
        status: target.status,
        attempts: target.attempts,
        nextAttemptAt: target.nextAttemptAt,
        lastError: target.lastError,
      };
    }),
    auditLogs: logs.map((log) => ({
      id: log.id,
      actorUsername: usernames.get(log.ownerUserId) ?? "Unknown",
      level: log.level,
      event: log.event,
      message: log.message,
      ip: log.metadata && typeof log.metadata === "object" && "ip" in log.metadata && typeof log.metadata.ip === "string" ? log.metadata.ip : null,
      campaignId: log.campaignId,
      accountId: log.accountId,
      createdAt: log.createdAt,
    })),
  }));
});

router.patch("/admin/campaigns/:campaignId", async (req, res): Promise<void> => {
  const params = UpdateAdminCampaignStatusParams.safeParse(req.params);
  const body = UpdateAdminCampaignStatusBody.safeParse(req.body);
  if (!params.success || !body.success) return void sendError(res, 400, "Trạng thái campaign không hợp lệ.");
  const [campaign] = await db.update(campaignsTable).set({ status: body.data.status, updatedAt: new Date() })
    .where(eq(campaignsTable.id, params.data.campaignId))
    .returning();
  if (!campaign) return void sendError(res, 404, "Không tìm thấy campaign.");
  const targets = await db.select({ status: campaignTargetsTable.status }).from(campaignTargetsTable)
    .where(eq(campaignTargetsTable.campaignId, campaign.id));
  const [owner] = await db.select({ username: appUsersTable.username }).from(appUsersTable)
    .where(eq(appUsersTable.id, campaign.ownerUserId)).limit(1);
  await recordActivity({
    ownerUserId: req.userId!,
    event: "campaign.admin_status_updated",
    level: "success",
    campaignId: campaign.id,
    message: `Admin changed campaign status to ${body.data.status}`,
    metadata: { targetOwnerUserId: campaign.ownerUserId, status: body.data.status },
  });
  res.json(UpdateAdminCampaignStatusResponse.parse({
    id: campaign.id,
    ownerUsername: owner?.username ?? "Unknown",
    name: campaign.name,
    status: campaign.status,
    pendingTargets: targets.filter((target) => ["pending", "sending"].includes(target.status)).length,
    failedTargets: targets.filter((target) => target.status === "failed").length,
    sentTargets: targets.filter((target) => target.status === "sent").length,
  }));
});

router.post("/admin/targets/:targetId/retry", async (req, res): Promise<void> => {
  const params = RetryAdminCampaignTargetParams.safeParse(req.params);
  if (!params.success) return void sendError(res, 400, "Target không hợp lệ.");
  const [target] = await db.update(campaignTargetsTable).set({
    status: "pending",
    nextAttemptAt: new Date(),
    lastError: null,
    updatedAt: new Date(),
  }).where(and(
    eq(campaignTargetsTable.id, params.data.targetId),
    inArray(campaignTargetsTable.status, ["failed", "requires_review"]),
  )).returning();
  if (!target) {
    const [existing] = await db.select({ id: campaignTargetsTable.id }).from(campaignTargetsTable)
      .where(eq(campaignTargetsTable.id, params.data.targetId)).limit(1);
    return void sendError(res, existing ? 409 : 404, existing
      ? "Chỉ có thể retry target đang lỗi hoặc cần xem xét."
      : "Không tìm thấy target.");
  }
  const [campaign] = await db.select().from(campaignsTable).where(eq(campaignsTable.id, target.campaignId)).limit(1);
  if (campaign) {
    await db.update(campaignsTable).set({ status: "queued", updatedAt: new Date() })
      .where(eq(campaignsTable.id, campaign.id));
  }
  const [destination] = await db.select({ title: destinationsTable.title }).from(destinationsTable)
    .where(eq(destinationsTable.id, target.destinationId)).limit(1);
  const [owner] = campaign
    ? await db.select({ username: appUsersTable.username }).from(appUsersTable).where(eq(appUsersTable.id, campaign.ownerUserId)).limit(1)
    : [];
  await recordActivity({
    ownerUserId: req.userId!,
    event: "campaign.target.admin_retried",
    level: "success",
    campaignId: target.campaignId,
    targetId: target.id,
    message: "Admin requeued a campaign target",
    metadata: { targetOwnerUserId: campaign?.ownerUserId ?? null },
  });
  res.json(RetryAdminCampaignTargetResponse.parse({
    id: target.id,
    campaignId: target.campaignId,
    campaignName: campaign?.name ?? "Unknown campaign",
    ownerUsername: owner?.username ?? "Unknown",
    destinationTitle: destination?.title ?? null,
    status: target.status,
    attempts: target.attempts,
    nextAttemptAt: target.nextAttemptAt,
    lastError: target.lastError,
  }));
});

router.get("/admin/license-keys", async (req, res): Promise<void> => {
  const parsed = ListAdminLicenseKeysQueryParams.safeParse(req.query);
  if (!parsed.success) return void sendError(res, 400, "Bộ lọc license key không hợp lệ.");
  const licenses = await listAdminLicenseKeys(parsed.data);
  res.json(ListAdminLicenseKeysResponse.parse(licenses));
});

router.post("/admin/license-keys", async (req, res): Promise<void> => {
  const parsed = CreateAdminLicenseKeyBody.safeParse(req.body);
  if (!parsed.success) return void sendError(res, 400, "Thông tin license key không hợp lệ.");
  if (!Number.isInteger(parsed.data.durationDays) || !Number.isInteger(parsed.data.quantity)) {
    return void sendError(res, 400, "Thời hạn và số lượng license key phải là số nguyên.");
  }
  try {
    const created = await createAdminLicenseKeys({
      ...parsed.data,
      createdBy: req.userId!,
      createdByUsername: req.authUser!.username,
    });
    res.status(201).json(CreateAdminLicenseKeyResponse.parse(created));
  } catch (error) {
    req.log.error({ err: error }, "Unable to create license key");
    sendError(res, 500, "Không thể tạo license key lúc này.");
  }
});

router.post("/admin/license-keys/:licenseKeyId/revoke", async (req, res): Promise<void> => {
  const parsed = RevokeAdminLicenseKeyParams.safeParse(req.params);
  if (!parsed.success) return void sendError(res, 400, "License key không hợp lệ.");
  const outcome = await revokeAdminLicenseKey(parsed.data.licenseKeyId, req.userId!);
  if (outcome === "not_found") return void sendError(res, 404, "Không tìm thấy license key.");
  if (outcome === "not_available") {
    return void sendError(res, 409, "Chỉ có thể vô hiệu hóa key chưa được kích hoạt.");
  }
  res.status(204).send();
});

export default router;