# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/purchase-settings:
    get:
      operationId: getAdminPurchaseSettings
      tags: [admin]
      responses:
        "200":
          description: Admin-configured purchase links
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseSettings" }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
    patch:
      operationId: updateAdminPurchaseSettings
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseSettingsInput" }
      responses:
        "200":
          description: Purchase links updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseSettings" }
        "400": { description: Invalid Telegram purchase link }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
  /admin/system-settings:
    get:
      operationId: getAdminSystemSettings
      tags: [admin]
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSystemSettings" }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
    patch:
      operationId: updateAdminSystemSettings
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSystemSettingsInput" }
      responses:
        "200":
          description: System settings updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSystemSettings" }
        "400": { description: Invalid system settings }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
  /admin/operations:
    get:
      operationId: getAdminOperations
      tags: [admin]
      responses:
        "200":
          description: Operational monitoring data for administrators
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOperations" }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
  /admin/campaigns/{campaignId}:
    patch:
      operationId: updateAdminCampaignStatus
      tags: [admin]
      parameters:
        - { name: campaignId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminCampaignStatusInput" }
      responses:
        "200":
          description: Campaign status updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOperationCampaign" }
        "400": { description: Invalid status }
        "404": { description: Campaign not found }
  /admin/targets/{targetId}/retry:
    post:
      operationId: retryAdminCampaignTarget
      tags: [admin]
      parameters:
        - { name: targetId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Campaign target requeued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOperationTarget" }
        "404": { description: Target not found }
...
  /admin/license-keys:
    get:
      operationId: listAdminLicenseKeys
      tags: [admin]
      parameters:
        - { name: status, in: query, required: false, schema: { $ref: "#/components/schemas/LicenseKeyStatus" } }
        - { name: plan, in: query, required: false, schema: { $ref: "#/components/schemas/PlanCode" } }
      responses:
        "200":
          description: License key inventory
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminLicenseKey" }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
    post:
      operationId: createAdminLicenseKey
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAdminLicenseKeyInput" }
      responses:
        "201":
          description: License key created; plaintext key is returned only in this response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateAdminLicenseKeyResult" }
        "400": { description: Invalid key parameters }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
  /admin/license-keys/{licenseKeyId}/revoke:
    post:
      operationId: revokeAdminLicenseKey
      tags: [admin]
      parameters:
        - { $ref: "#/components/parameters/LicenseKeyId" }
      responses:
        "204": { description: Unused license key revoked }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
        "404": { description: License key not found }
        "409": { description: License key is already claimed or revoked }
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      responses:
        "200":
          description: System-wide admin dashboard metrics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOverview" }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      parameters:
        - { name: search, in: query, required: false, schema: { type: string, maxLength: 64 } }
        - { name: plan, in: query, required: false, schema: { $ref: "#/components/schemas/PlanCode" } }
      responses:
        "200":
          description: Users with subscription and resource usage summaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      parameters:
        - { $ref: "#/components/parameters/AdminUserId" }
      responses:
        "200":
          description: User details for administrator support
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
        "404": { description: User not found }
  /admin/users/{userId}/subscription:
    patch:
      operationId: updateAdminUserSubscription
      tags: [admin]
      parameters:
        - { $ref: "#/components/parameters/AdminUserId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSubscriptionUpdateInput" }
      responses:
        "200":
          description: Subscription extended or upgraded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "400": { description: Invalid subscription change }
        "401": { description: Authentication is required }
        "403": { description: Administrator permission is required }
        "404": { description: User not found }
        "409": { description: Subscription cannot be downgraded }
components:
  parameters:
...
    LicenseKeyId:
      name: licenseKeyId
      in: path
      required: true
      schema: { type: string }
    AdminUserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    PlanCode:
      type: string
      enum: [plus, pro, unlimited]
...
    Subscription:
      type: object
      required: [plan, startedAt, expiresAt, status, accountLimit, campaignLimit, messageDailyLimit]
      properties:
        plan: { $ref: "#/components/schemas/PlanCode" }
        startedAt: { type: string, format: date-time }
        expiresAt: { type: ["string", "null"], format: date-time }
        status: { type: string, enum: [active, expired] }
        accountLimit: { type: ["number", "null"] }
        campaignLimit: { type: ["number", "null"] }
        messageDailyLimit: { type: ["number", "null"] }
...
    PurchaseSettings:
      type: object
      required: [telegramPurchaseUrl]
      properties:
        telegramPurchaseUrl: { type: ["string", "null"] }
    PurchaseSettingsInput:
      type: object
      required: [telegramPurchaseUrl]
      properties:
        telegramPurchaseUrl: { type: string, minLength: 1, maxLength: 512 }
    PlanLimitSettings:
      type: object
      required: [accountLimit, campaignLimit, messageDailyLimit]
      properties:
        accountLimit: { type: ["number", "null"], minimum: 0 }
        campaignLimit: { type: ["number", "null"], minimum: 0 }
        messageDailyLimit: { type: ["number", "null"], minimum: 0 }
    CampaignDefaults:
      type: object
      required: [maxRetries, delayMinSeconds, delayMaxSeconds, roundDelayMinSeconds, roundDelayMaxSeconds]
      properties:
        maxRetries: { type: number, minimum: 0, maximum: 20 }
        delayMinSeconds: { type: number, minimum: 0, maximum: 120 }
        delayMaxSeconds: { type: number, minimum: 0, maximum: 120 }
        roundDelayMinSeconds: { type: number, minimum: 0, maximum: 259200 }
        roundDelayMaxSeconds: { type: number, minimum: 0, maximum: 259200 }
...
    AdminSystemSettings:
      type: object
      required: [planLimits, defaultAccountDailyLimit, campaignDefaults, registrationEnabled, maintenanceMode, defaultTimezone]
      properties:
        planLimits:
          type: object
          required: [plus, pro, unlimited]
          properties:
            plus: { $ref: "#/components/schemas/PlanLimitSettings" }
            pro: { $ref: "#/components/schemas/PlanLimitSettings" }
            unlimited: { $ref: "#/components/schemas/PlanLimitSettings" }
        defaultAccountDailyLimit: { type: number, minimum: 1, maximum: 100000 }
        campaignDefaults: { $ref: "#/components/schemas/CampaignDefaults" }
        registrationEnabled: { type: boolean }
        maintenanceMode: { type: boolean }
        defaultTimezone: { type: string, minLength: 1, maxLength: 100 }
    AdminSystemSettingsInput:
      type: object
      required: [planLimits, defaultAccountDailyLimit, campaignDefaults, registrationEnabled, maintenanceMode, defaultTimezone]
      properties:
        planLimits:
          type: object
          required: [plus, pro, unlimited]
          properties:
            plus: { $ref: "#/components/schemas/PlanLimitSettings" }
            pro: { $ref: "#/components/schemas/PlanLimitSettings" }
            unlimited: { $ref: "#/components/schemas/PlanLimitSettings" }
        defaultAccountDailyLimit: { type: number, minimum: 1, maximum: 100000 }
        campaignDefaults: { $ref: "#/components/schemas/CampaignDefaults" }
        registrationEnabled: { type: boolean }
        maintenanceMode: { type: boolean }
        defaultTimezone: { type: string, minLength: 1, maxLength: 100 }
    AdminOperationAccount:
      type: object
      required: [id, ownerUsername, name, status, proxyName, proxyStatus, lastSyncAt, cooldownUntil, failedTargets]
      properties:
        id: { type: string }
        ownerUsername: { type: string }
        name: { type: string }
        status: { type: string }
        proxyName: { type: ["string", "null"] }
        proxyStatus: { type: ["string", "null"] }
        lastSyncAt: { type: ["string", "null"], format: date-time }
        cooldownUntil: { type: ["string", "null"], format: date-time }
        failedTargets: { type: number }
    AdminOperationCampaign:
      type: object
      required: [id, ownerUsername, name, status, pendingTargets, failedTargets, sentTargets]
      properties:
        id: { type: string }
        ownerUsername: { type: string }
        name: { type: string }
        status: { type: string }
        pendingTargets: { type: number }
        failedTargets: { type: number }
        sentTargets: { type: number }
    AdminCampaignStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [queued, paused] }
    AdminOperationTarget:
      type: object
      required: [id, campaignId, campaignName, ownerUsername, destinationTitle, status, attempts, nextAttemptAt, lastError]
      properties:
        id: { type: string }
        campaignId: { type: string }
        campaignName: { type: string }
        ownerUsername: { type: string }
        destinationTitle: { type: ["string", "null"] }
        status: { type: string }
        attempts: { type: number }
        nextAttemptAt: { type: ["string", "null"], format: date-time }
        lastError: { type: ["string", "null"] }
    AdminAuditLog:
      type: object
      required: [id, actorUsername, level, event, message, ip, campaignId, accountId, createdAt]
      properties:
        id: { type: string }
        actorUsername: { type: string }
        level: { type: string }
        event: { type: string }
        message: { type: string }
        ip: { type: ["string", "null"] }
        campaignId: { type: ["string", "null"] }
        accountId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    AdminOperations:
      type: object
      required: [accounts, campaigns, targets, auditLogs]
      properties:
        accounts:
          type: array
          items: { $ref: "#/components/schemas/AdminOperationAccount" }
        campaigns:
          type: array
          items: { $ref: "#/components/schemas/AdminOperationCampaign" }
        targets:
          type: array
          items: { $ref: "#/components/schemas/AdminOperationTarget" }
        auditLogs:
          type: array
          items: { $ref: "#/components/schemas/AdminAuditLog" }
...
    LicenseKeyStatus:
      type: string
      enum: [available, claimed, revoked]
    AdminLicenseKey:
      type: object
      required: [id, plan, durationDays, label, status, createdAt, createdByUsername, claimedAt, claimedByUsername, revokedAt, revokedByUsername]
      properties:
        id: { type: string }
        plan: { $ref: "#/components/schemas/PlanCode" }
        durationDays: { type: number }
        label: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/LicenseKeyStatus" }
        createdAt: { type: string, format: date-time }
        createdByUsername: { type: ["string", "null"] }
        claimedAt: { type: ["string", "null"], format: date-time }
        claimedByUsername: { type: ["string", "null"] }
        revokedAt: { type: ["string", "null"], format: date-time }
        revokedByUsername: { type: ["string", "null"] }
    CreateAdminLicenseKeyInput:
      type: object
      required: [plan, durationDays, quantity]
      properties:
        plan: { type: string, enum: [pro, unlimited] }
        durationDays: { type: number, minimum: 1, maximum: 3660, multipleOf: 1 }
        quantity: { type: number, minimum: 1, maximum: 100, multipleOf: 1 }
        label: { type: string, minLength: 1, maxLength: 120 }
    CreateAdminLicenseKeyResult:
      type: object
      required: [licenseKeys, licenses]
      properties:
        licenseKeys:
          type: array
          items: { type: string }
        licenses:
          type: array
          items: { $ref: "#/components/schemas/AdminLicenseKey" }
    AdminOverview:
      type: object
      required: [usersTotal, usersNewLast30Days, usersAdmins, subscriptions, licenses, telegramAccountsTotal, telegramAccountsConnected, campaignsTotal, campaignsQueued, campaignsFailed]
      properties:
        usersTotal: { type: number }
        usersNewLast30Days: { type: number }
        usersAdmins: { type: number }
        subscriptions:
          $ref: "#/components/schemas/AdminPlanDistribution"
        licenses:
          $ref: "#/components/schemas/AdminLicenseDistribution"
        telegramAccountsTotal: { type: number }
        telegramAccountsConnected: { type: number }
        campaignsTotal: { type: number }
        campaignsQueued: { type: number }
        campaignsFailed: { type: number }
    AdminPlanDistribution:
      type: object
      required: [plus, pro, unlimited, expired]
      properties:
        plus: { type: number }
        pro: { type: number }
        unlimited: { type: number }
        expired: { type: number }
    AdminLicenseDistribution:
      type: object
      required: [available, claimed, revoked]
      properties:
        available: { type: number }
        claimed: { type: number }
        revoked: { type: number }
...
    AdminUser:
      type: object
      required: [id, username, role, joinedAt, lastActiveAt, storedPlan, subscription, usage]
      properties:
        id: { type: string }
        username: { type: string }
        role: { type: string, enum: [user, admin] }
        joinedAt: { type: string, format: date-time }
        lastActiveAt: { type: ["string", "null"], format: date-time }
        storedPlan: { $ref: "#/components/schemas/PlanCode" }
        subscription: { $ref: "#/components/schemas/Subscription" }
        usage: { $ref: "#/components/schemas/AdminUserUsage" }
    AdminSubscriptionUpdateInput:
      type: object
      required: [plan, durationDays]
      properties:
        plan: { $ref: "#/components/schemas/PlanCode" }
        durationDays: { type: number, minimum: 1, maximum: 3660 }
...